requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.9.2

# LangChain & RAG
langchain==0.3.2
langchain-community==0.3.2
langchain-core==0.3.9
langchain-text-splitters==0.3.2

# Vector DB
chromadb==0.5.7

# Embeddings (local + API)
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.0

# Retrieval
rank-bm25==0.2.2
numpy==1.26.4

# Content Extraction
feedparser==6.0.11
beautifulsoup4==4.12.3
trafilatura==1.12.2
youtube-transcript-api==0.6.2
podcastparser==0.6.10

# Whisper (for podcast transcription)
openai-whisper==20231117

# API
fastapi==0.115.2
uvicorn==0.31.1
httpx==0.27.2

# Evaluation
ragas==0.2.0
datasets==3.0.1

# Optional API upgrades (Phase 2)
anthropic==0.34.2
voyageai==0.2.3
cohere==5.9.4
qdrant-client==1.11.3

# Monitoring
langfuse==2.51.0
